Created integration tests.
These integration tests dynamically load any JSON files in testdata,
dynamically mimics its Kartograph mapping. Then it uses the generated
mapping to regenerate the JSON. Then it compares with the original file.

I used this to also compare allocation stack. So if stackprof is installed
and temporarily in the Gemfile, it will load and generate the profile for
every json file in test data. This is useful to trace how many objects
Kartograph allocates during the generation of the JSON dictionaries.